HG#65 update to laravel 8 and refactor factories to match the new style.
