+++ app/Models/User.php
@@ -2,6 +2,7 @@
 
 namespace App\Models;
 
+use Illuminate\Database\Eloquent\Factories\HasFactory;
 use Illuminate\Database\Eloquent\SoftDeletes;
 use Illuminate\Foundation\Auth\User as Authenticatable;
 use Illuminate\Notifications\Notifiable;
@@ -13,7 +14,7 @@ use Illuminate\Notifications\Notifiable;
  */
 class User extends Authenticatable
 {
-    use Notifiable, SoftDeletes;
+    use Notifiable, SoftDeletes, HasFactory;
 
     /**
      * The attributes that are mass assignable.
@@ -21,12 +22,12 @@ class User extends Authenticatable
      * @var array
      */
     protected $fillable
-        = [
-            'first_name',
-            'last_name',
-            'email',
-            'password',
-        ];
+            = [
+                    'first_name',
+                    'last_name',
+                    'email',
+                    'password',
+            ];
 
     /**
      * The attributes that should be hidden for arrays.
@@ -34,10 +35,10 @@ class User extends Authenticatable
      * @var array
      */
     protected $hidden
-        = [
-            'password',
-            'remember_token',
-        ];
+            = [
+                    'password',
+                    'remember_token',
+            ];
 
     public function getNameAttribute()
     {
