+++ composer.json
@@ -13,13 +13,13 @@
     "ext-json": "*",
     "fideloper/proxy": ">=4 <5.0.0",
     "fruitcake/laravel-cors": "^1.0",
-    "guzzlehttp/guzzle": ">=6.3 <7.0.0",
+    "guzzlehttp/guzzle": "^7.0.1",
     "intervention/image": ">=2.5 <3.0.0",
-    "laravel/framework": ">=7.0 <8.0.0",
+    "laravel/framework": "^8.0",
     "laravel/helpers": "^1.2",
     "laravel/tinker": "^2.0",
-    "laravel/ui": "^2.0",
-    "spatie/laravel-google-calendar": "^2.5"
+    "laravel/ui": "^3.0",
+    "spatie/laravel-google-calendar": "^3.0"
   },
   "require-dev": {
     "ext-sqlite3": "^7.2",
@@ -29,12 +29,12 @@
     "filp/whoops": "^2.7",
     "fzaninotto/faker": "^1.9",
     "mockery/mockery": "^1.3",
-    "nunomaduro/collision": "^4.1",
-    "phpunit/phpunit": "^8.5"
+    "nunomaduro/collision": "^5.0",
+    "phpunit/phpunit": "^9.0"
   },
   "autoload": {
     "classmap": [
-      "database/seeds",
+      "database/seeders",
       "database/factories"
     ],
     "psr-4": {
