--- database/factories/ModelsOrgFactory.php
@@ -1,9 +1,0 @@
-<?php
-
-use Faker\Generator as Faker;
-
-$factory->define(App\Models\Org::class, function (Faker $faker) {
-    return [
-        //
-    ];
-});
