--- database/seeds/UsersTableSeeder.php
@@ -1,21 +1,0 @@
-<?php
-
-use Illuminate\Database\Seeder;
-
-class UsersTableSeeder extends Seeder
-{
-    /**
-     * Run the database seeds.
-     *
-     * @return void
-     */
-    public function run()
-    {
-        \App\Models\User::firstOrCreate(['email' => 'admin@admin.com'], [
-            'first_name' => 'admin',
-            'last_name'  => 'admin',
-            'email'      => 'admin@admin.com',
-            'password'   => bcrypt('admin'),
-        ]);
-    }
-}
