+++ package.json
@@ -1,53 +1,53 @@
 {
-    "name": "hackgreenville_com",
-    "license": "SEE LICENSE IN LICENSE.MD",
-    "description": "website for hackgreenville",
-    "private": false,
-    "scripts": {
-      "dev": "npm run development",
-      "development": "cross-env NODE_ENV=development node_modules/webpack/bin/webpack.js --progress --hide-modules --config=node_modules/laravel-mix/setup/webpack.config.js",
-      "watch": "npm run development -- --watch",
-      "watch-poll": "npm run watch -- --watch-poll",
-      "hot": "cross-env NODE_ENV=development node_modules/webpack-dev-server/bin/webpack-dev-server.js --inline --hot --config=node_modules/laravel-mix/setup/webpack.config.js",
-      "prod": "npm run production",
-      "test": "echo 'Running art test'",
-      "posttest": "php artisan test",
-      "production": "cross-env NODE_ENV=production node_modules/webpack/bin/webpack.js --no-progress --hide-modules --config=node_modules/laravel-mix/setup/webpack.config.js",
-      "hg:install": "yarn install; composer install",
-      "hg:deploy-prep": "yarn hg:install; yarn production",
-      "hg:watch": "pm2 delete hg:watch; pm2 start --only='hg:watch'",
-      "hg:serve": "pm2 delete hg:serve; pm2 start --only='hg:serve'",
-      "hg:queue": "pm2 delete hg:queue; pm2 start --only='hg:queue'",
-      "hg:monit": "pm2 monit",
-      "hg:delete": "pm2 delete hg:watch hg:serve hg:queue",
-      "hg:dev": "yarn hg:delete; yarn hg:watch && yarn hg:serve",
-      "hg:dev-enable-queue": "yarn hg:delete; yarn hg:watch && yarn hg:serve && yarn hg:queue"
-    },
-    "devDependencies": {
-      "axios": "^0.19.2",
-      "bootstrap": "^4.0.0",
-      "browser-sync": "^2.26.7",
-      "browser-sync-webpack-plugin": "2.2.2",
-      "cross-env": "^7.0",
-      "jquery": "^3.2",
-      "laravel-mix": "^5.0.4",
-      "lodash": "^4.17.15",
-      "popper.js": "^1.12",
-      "resolve-url-loader": "^2.3.1",
-      "sass": "^1.20.1",
-      "sass-loader": "^8.0.0",
-      "sharrre": "^2.0.1",
-      "vue": "^2.5.17",
-      "vue-template-compiler": "^2.6.10"
-    },
-    "dependencies": {
-        "@fullcalendar/core": "^4.4.0",
-        "@fullcalendar/daygrid": "^4.4.0",
-        "bootstrap-datepicker": "^1.9.0",
-        "bootstrap-switch": "^3.4.0",
-        "browsersync": "^0.0.1-security",
-        "dotenv": "latest",
-        "moment": "^2.24.0",
-        "sweetalert2": "^9.10.6"
-    }
+  "name": "hackgreenville_com",
+  "license": "SEE LICENSE IN LICENSE.MD",
+  "description": "website for hackgreenville",
+  "private": false,
+  "scripts": {
+    "dev": "npm run development",
+    "development": "cross-env NODE_ENV=development node_modules/webpack/bin/webpack.js --progress --hide-modules --config=node_modules/laravel-mix/setup/webpack.config.js",
+    "watch": "npm run development -- --watch",
+    "watch-poll": "npm run watch -- --watch-poll",
+    "hot": "cross-env NODE_ENV=development node_modules/webpack-dev-server/bin/webpack-dev-server.js --inline --hot --config=node_modules/laravel-mix/setup/webpack.config.js",
+    "prod": "npm run production",
+    "test": "echo 'Running art test'",
+    "posttest": "php artisan test",
+    "production": "cross-env NODE_ENV=production node_modules/webpack/bin/webpack.js --no-progress --hide-modules --config=node_modules/laravel-mix/setup/webpack.config.js",
+    "hg:install": "yarn install; composer install",
+    "hg:deploy-prep": "yarn hg:install; yarn production",
+    "hg:watch": "pm2 delete hg:watch; pm2 start --only='hg:watch'",
+    "hg:serve": "pm2 delete hg:serve; pm2 start --only='hg:serve'",
+    "hg:queue": "pm2 delete hg:queue; pm2 start --only='hg:queue'",
+    "hg:monit": "pm2 monit",
+    "hg:delete": "pm2 delete hg:watch hg:serve hg:queue",
+    "hg:dev": "yarn hg:delete; yarn hg:watch && yarn hg:serve",
+    "hg:dev-enable-queue": "yarn hg:delete; yarn hg:watch && yarn hg:serve && yarn hg:queue"
+  },
+  "devDependencies": {
+    "axios": "^0.19.2",
+    "bootstrap": "^4.0.0",
+    "browser-sync": "^2.26.7",
+    "browser-sync-webpack-plugin": "2.2.2",
+    "cross-env": "^7.0",
+    "jquery": "^3.2",
+    "laravel-mix": "^5.0.4",
+    "lodash": "^4.17.15",
+    "popper.js": "^1.12",
+    "resolve-url-loader": "^2.3.1",
+    "sass": "^1.20.1",
+    "sass-loader": "^8.0.0",
+    "sharrre": "^2.0.1",
+    "vue": "^2.5.17",
+    "vue-template-compiler": "^2.6.10"
+  },
+  "dependencies": {
+    "@fullcalendar/core": "^4.4.0",
+    "@fullcalendar/daygrid": "^4.4.0",
+    "bootstrap-datepicker": "^1.9.0",
+    "bootstrap-switch": "^3.4.0",
+    "browsersync": "^0.0.1-security",
+    "dotenv": "latest",
+    "moment": "^2.24.0",
+    "sweetalert2": "^9.10.6"
+  }
 }
